+++ typescript-strict/index.js
@@ -1,6 +1,11 @@
 module.exports = {
   rules: {
-    '@typescript-eslint/explicit-function-return-type': ['error'],
+    '@typescript-eslint/explicit-function-return-type': [
+      'error',
+      {
+        allowExpressions: true,
+      },
+    ],
     '@typescript-eslint/no-implicit-any-catch': ['error'],
     '@typescript-eslint/no-inferrable-types': ['warn'],
     '@typescript-eslint/no-non-null-assertion': ['error'],